Move decoder constants
