+++ src/ImageSharp/Formats/Jpeg/Components/Decoder/ColorConverters/JpegColorConverter.FromGrayScale.cs
@@ -4,23 +4,21 @@
 using System;
 using System.Numerics;
 
-namespace SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder.ColorConverters
+namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters
 {
     internal abstract partial class JpegColorConverter
     {
-        internal class FromRgb : ColorConverters.JpegColorConverter
+        internal class FromGrayscale : JpegColorConverter
         {
-            public FromRgb()
-                : base(JpegColorSpace.RGB)
+            public FromGrayscale()
+                : base(JpegColorSpace.Grayscale)
             {
             }
 
             public override void ConvertToRGBA(ComponentValues values, Span<Vector4> result)
             {
                 // TODO: We can optimize a lot here with Vector<float> and SRCS.Unsafe()!
-                ReadOnlySpan<float> rVals = values.Component0;
-                ReadOnlySpan<float> gVals = values.Component1;
-                ReadOnlySpan<float> bVals = values.Component2;
+                ReadOnlySpan<float> yVals = values.Component0;
 
                 var v = new Vector4(0, 0, 0, 1);
 
@@ -28,13 +26,11 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder.ColorConverters
 
                 for (int i = 0; i < result.Length; i++)
                 {
-                    float r = rVals[i];
-                    float g = gVals[i];
-                    float b = bVals[i];
+                    float y = yVals[i];
 
-                    v.X = r;
-                    v.Y = g;
-                    v.Z = b;
+                    v.X = y;
+                    v.Y = y;
+                    v.Z = y;
 
                     v *= scale;
