+++ src/ImageSharp/Formats/Jpeg/Components/Decoder/JpegColorSpace.cs
@@ -1,4 +1,7 @@
-namespace SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder
+// Copyright (c) Six Labors and contributors.
+// Licensed under the Apache License, Version 2.0.
+
+namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder
 {
     /// <summary>
     /// Identifies the colorspace of a Jpeg image
