+++ src/ImageSharp/Formats/Jpeg/Components/Decoder/ProfileResolver.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Text;
 
-namespace SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder
+namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder
 {
     /// <summary>
     /// Provides methods for identifying metadata and color profiles within jpeg images.
