+++ tests/ImageSharp.Tests/Formats/Jpg/JFifMarkerTests.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder;
+using SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder;
 
 using Xunit;
 
@@ -27,7 +27,7 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
         [Fact]
         public void MarkerReturnsCorrectParsedValue()
         {
-            bool isJFif = JFifMarker.TryParse(this.bytes, out var marker);
+            bool isJFif = JFifMarker.TryParse(this.bytes, out JFifMarker marker);
 
             Assert.True(isJFif);
             Assert.Equal(1, marker.MajorVersion);
@@ -40,26 +40,26 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
         [Fact]
         public void MarkerIgnoresIncorrectValue()
         {
-            bool isJFif = JFifMarker.TryParse(new byte[] { 0, 0, 0, 0 }, out var marker);
+            bool isJFif = JFifMarker.TryParse(new byte[] { 0, 0, 0, 0 }, out JFifMarker marker);
 
             Assert.False(isJFif);
-            Assert.Equal(default(JFifMarker), marker);
+            Assert.Equal(default, marker);
         }
 
         [Fact]
         public void MarkerIgnoresCorrectHeaderButInvalidDensities()
         {
-            bool isJFif = JFifMarker.TryParse(this.bytes3, out var marker);
+            bool isJFif = JFifMarker.TryParse(this.bytes3, out JFifMarker marker);
 
             Assert.False(isJFif);
-            Assert.Equal(default(JFifMarker), marker);
+            Assert.Equal(default, marker);
         }
 
         [Fact]
         public void MarkerEqualityIsCorrect()
         {
-            JFifMarker.TryParse(this.bytes, out var marker);
-            JFifMarker.TryParse(this.bytes, out var marker2);
+            JFifMarker.TryParse(this.bytes, out JFifMarker marker);
+            JFifMarker.TryParse(this.bytes, out JFifMarker marker2);
 
             Assert.True(marker.Equals(marker2));
         }
@@ -67,8 +67,8 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
         [Fact]
         public void MarkerInEqualityIsCorrect()
         {
-            JFifMarker.TryParse(this.bytes, out var marker);
-            JFifMarker.TryParse(this.bytes2, out var marker2);
+            JFifMarker.TryParse(this.bytes, out JFifMarker marker);
+            JFifMarker.TryParse(this.bytes2, out JFifMarker marker2);
 
             Assert.False(marker.Equals(marker2));
         }
@@ -76,8 +76,8 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
         [Fact]
         public void MarkerHashCodeIsReplicable()
         {
-            JFifMarker.TryParse(this.bytes, out var marker);
-            JFifMarker.TryParse(this.bytes, out var marker2);
+            JFifMarker.TryParse(this.bytes, out JFifMarker marker);
+            JFifMarker.TryParse(this.bytes, out JFifMarker marker2);
 
             Assert.True(marker.GetHashCode().Equals(marker2.GetHashCode()));
         }
@@ -85,8 +85,8 @@ namespace SixLabors.ImageSharp.Tests.Formats.Jpg
         [Fact]
         public void MarkerHashCodeIsUnique()
         {
-            JFifMarker.TryParse(this.bytes, out var marker);
-            JFifMarker.TryParse(this.bytes2, out var marker2);
+            JFifMarker.TryParse(this.bytes, out JFifMarker marker);
+            JFifMarker.TryParse(this.bytes2, out JFifMarker marker2);
 
             Assert.False(marker.GetHashCode().Equals(marker2.GetHashCode()));
         }
