+++ tests/ImageSharp.Tests/Formats/Jpg/JpegColorConverterTests.cs
@@ -6,8 +6,8 @@ using System.Numerics;
 
 using SixLabors.ImageSharp.ColorSpaces;
 using SixLabors.ImageSharp.ColorSpaces.Conversion;
-using SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder;
-using SixLabors.ImageSharp.Formats.Jpeg.Common.Decoder.ColorConverters;
+using SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder;
+using SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder.ColorConverters;
 using SixLabors.ImageSharp.Memory;
 
 using Xunit;
